# backend/api.yaml
openapi: "3.0.3"
info:
  title: Skill Explorers - Life Skills App
  version: 1.0.0
  description: |
    ## About the Application
    Skill Explorers is a Life Skills App designed for school-aged children, parents, and teachers. 
    It fosters emotional well-being, daily responsibility, creativity, and cognitive skill-building 
    through journaling, gamified tasks, infotainment, and story-based learning. 


    ## API and User Story Mapping

    | Endpoint                                      | Description                     | User Story                      |
    |-----------------------------------------------|---------------------------------|---------------------------------|
    | `/todo`                                       | Child’s to-do management        | Primary User: To-do             |
    | `/generate_story`                             | AI Story generation via Gemini  | Primary User: Story             |
    | `/submit_quiz`                                | Auto-quiz generation from story | Tertiary User: Fun quiz         |
    | `/journal`                                    | Write daily journal             | Primary User: Journaling        |
    | `/infotainment/generate`                      | AI infotainment (Gemini)        |Secondary User: Infotainment     |
    | `/calendar-report`                            | Daily completion colors         | Primary User: Streak            |
...

    ---
    **GenAI Highlights**:
...
    ---

    **External Integrations**
...

...

...
servers:
...
components:
  securitySchemes:
    bearerAuthParent:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bearerAuthChild:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /register:
    post:
      tags:
        - Authentication
      summary: Register a new parent user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
                - username
              properties:
                email:
                  type: string
                  format: email
                  example: parent@example.com
                password:
                  type: string
                  example: mysecurepassword
                name:
                  type: string
                  example: John Doe
                username:
                  type: string
                  example: johndoe123
      responses:
        "201":
          description: Parent registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Parent registered successfully
        "400":
          description: Bad request due to missing fields or duplicate email/username
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Email already exists
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Something went wrong
  /login:
    post:
      tags:
        - Authentication
      summary: Child/Parent login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: johndoe123
                password:
                  type: string
                  example: mysecurepassword
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  access_token:
                    type: string
                    example: eyJ0eXAiOiJKV1QiLCJh...
                  role:
                    type: string
                    enum: [parent, child]
                    example: parent
                  redirect_to:
                    type: string
                    example: /parent_dashboard
        "400":
          description: Missing username or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Username and password are required
        "401":
          description: Invalid username or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid username, password, or role
        "415":
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Request content type must be application/json
  /add-child:
    post:
      tags:
        - Parent Actions
      summary: Add a new child user (parent only)
      description: Adds a child account under the authenticated parent. Requires JWT token with `parent` role.
      security:
        - bearerAuthParent: [] # Assumes JWT-based bearer authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - name
                - age
              properties:
                username:
                  type: string
                  example: child_user1
                password:
                  type: string
                  example: securechildpassword
                name:
                  type: string
                  example: Alice
                age:
                  type: integer
                  example: 10
                gender:
                  type: string
                  enum: [male, female, other]
                  example: female
      responses:
        "201":
          description: Child added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Child added successfully
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: All fields (username, password, name, age) are required.
        "401":
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Missing Authorization Header
        "404":
          description: Parent not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Parent not found
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Username already taken
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: SOmething wrong
...